Move the webpacker config override to @decidim/webpacker
