+++ packages/webpacker/index.js
@@ -1,0 +1,6 @@
+const webpacker = require("@rails/webpacker");
+const overrideConfig = require("./src/override-config");
+
+module.exports = Object.assign(webpacker, {
+  webpackConfig: overrideConfig(webpacker.webpackConfig)
+});
