+++ packages/webpacker/package.json
@@ -13,6 +13,7 @@
     "node": "^15.14.0",
     "npm": "^7.7.2"
   },
+  "main": "index.js",
   "dependencies": {
     "@babel/core": "^7.13.13",
     "@babel/eslint-parser": "^7.13.14",
